+++ lib/MapboxQueueProcessingFargateService.ts
@@ -1,16 +1,27 @@
 import { Construct } from 'constructs';
-import { FargateService, FargateTaskDefinition, Volume } from 'aws-cdk-lib/aws-ecs';
+import {
+  ContainerImage,
+  FargateService,
+  FargateTaskDefinition,
+  Volume
+} from 'aws-cdk-lib/aws-ecs';
 import {
   QueueProcessingFargateServiceProps,
   QueueProcessingServiceBase
 } from 'aws-cdk-lib/aws-ecs-patterns';
-import { cx_api, FeatureFlags } from 'aws-cdk-lib';
+import * as lambda from 'aws-cdk-lib/aws-lambda';
+import { cx_api, FeatureFlags, Duration } from 'aws-cdk-lib';
+import * as events from 'aws-cdk-lib/aws-events';
+import * as targets from 'aws-cdk-lib/aws-events-targets';
+import * as iam from 'aws-cdk-lib/aws-iam';
+import * as cloudwatch from 'aws-cdk-lib/aws-cloudwatch';
+import * as appscaling from 'aws-cdk-lib/aws-applicationautoscaling';
 
 /**
  * The properties for the MapboxQueueProcessingFargateService service.
  */
 export interface MapboxQueueProcessingFargateServiceProps
-  extends QueueProcessingFargateServiceProps {
+  extends Omit<QueueProcessingFargateServiceProps, 'image'>{
   /**
    * Specifies whether the container is marked as privileged. When this parameter is true, the container is given elevated privileges on the host container instance (similar to the root user).
    * @default false
@@ -34,6 +45,18 @@ export interface MapboxQueueProcessingFargateServiceProps
    * @default []
    */
   readonly volumes?: Volume[];
+
+  /**
+   * The ECS image to use here. We're overriding the default image property because it can be undefined
+   */
+  readonly image: ContainerImage;
+
+  /**
+   * The period between scale up events
+   * @default Duration.minutes(5)
+   */
+  readonly scaleUpAlarmInterval: Duration;
+
 }
 
 /**
@@ -51,6 +74,16 @@ export class MapboxQueueProcessingFargateService extends QueueProcessingServiceB
   public readonly taskDefinition: FargateTaskDefinition;
 
   /**
+   * A lambda to calculate the total messages (visible and not visible) in the SQS queue as a cloudwatch metric
+   */
+  readonly totalMessagesLambda: lambda.Function;
+
+  /**
+   * A metric to track the total messages visible and not visible
+   */
+  readonly totalMessagesMetric: cloudwatch.Metric;
+
+  /**
    * Constructs a new instance of the QueueProcessingFargateService class.
    */
   constructor(scope: Construct, id: string, props: MapboxQueueProcessingFargateServiceProps) {
@@ -60,6 +93,7 @@ export class MapboxQueueProcessingFargateService extends QueueProcessingServiceB
     this.taskDefinition = new FargateTaskDefinition(this, 'QueueProcessingTaskDef', {
       memoryLimitMiB: props.memoryLimitMiB || 512,
       cpu: props.cpu || 256,
+      ephemeralStorageGiB: props.ephemeralStorageGiB,
       family: props.family,
       runtimePlatform: props.runtimePlatform,
       volumes: props.volumes
@@ -76,7 +110,7 @@ export class MapboxQueueProcessingFargateService extends QueueProcessingServiceB
       healthCheck: props.healthCheck,
       privileged: props.privileged,
       memoryReservationMiB: props.memoryReservationMiB,
-      readonlyRootFilesystem: props.readonlyRootFilesystem
+      readonlyRootFilesystem: props.readonlyRootFilesystem,
     });
 
     // The desiredCount should be removed from the fargate service when the feature flag is removed.
@@ -105,7 +139,89 @@ export class MapboxQueueProcessingFargateService extends QueueProcessingServiceB
       enableExecuteCommand: props.enableExecuteCommand
     });
 
-    this.configureAutoscalingForService(this.service);
     this.grantPermissionsToService(this.service);
+
+    this.totalMessagesLambda = new lambda.Function(this, 'TotalMessagesLambda', {
+      runtime: lambda.Runtime.NODEJS_18_X,
+      handler: 'index.handler',
+      code: new lambda.InlineCode(`
+        const { SQS } = require('@aws-sdk/client-sqs');
+        const { CloudWatch } = require('@aws-sdk/client-cloudwatch');
+        exports.handler = function(event, context, callback) {
+          const sqs = new SQS({ region: process.env.AWS_DEFAULT_REGION });
+          const cw = new CloudWatch({ region: process.env.AWS_DEFAULT_REGION });
+
+          return sqs.getQueueAttributes({
+            QueueUrl: '${this.sqsQueue.queueUrl}',
+            AttributeNames: ['ApproximateNumberOfMessagesNotVisible', 'ApproximateNumberOfMessages']
+          })
+            .then((attrs) => {
+              return cw.putMetricData({
+                Namespace: 'Mapbox/ecs-watchbot',
+                MetricData: [{
+                  MetricName: 'TotalMessages',
+                  Dimensions: [{ Name: 'QueueName', Value: '${this.sqsQueue.queueName}'}],
+                  Value: Number(attrs.Attributes.ApproximateNumberOfMessagesNotVisible) +
+                          Number(attrs.Attributes.ApproximateNumberOfMessages)
+                }]
+              })
+            })
+            .then((metric) => callback(null, metric))
+            .catch((err) => callback(err));
+        }
+      `)
+    })
+
+    this.totalMessagesLambda.addToRolePolicy(new iam.PolicyStatement({
+      actions: ['cloudwatch:PutMetricData'],
+      resources: ['*']
+    }))
+    this.totalMessagesLambda.addToRolePolicy(new iam.PolicyStatement({
+      actions: ['sqs:GetQueueAttributes'],
+      resources: [this.sqsQueue.queueArn]
+    }))
+
+    const rule = new events.Rule(this, 'TotalMessagesRule', {
+      description: 'Update TotalMessages metric every minute',
+      schedule: events.Schedule.cron({ minute: '0/1'}) // run every minute
+    });
+
+    rule.addTarget(new targets.LambdaFunction(this.totalMessagesLambda));
+
+    this.totalMessagesMetric = new cloudwatch.Metric({
+      namespace: 'Mapbox/ecs-watchbot',
+      metricName: 'TotalMessages',
+      dimensionsMap: { QueueName: this.sqsQueue.queueName },
+      period: Duration.minutes(10),
+    });
+
+    const scalingTarget = new appscaling.ScalableTarget(this, 'WatchbotScalingTarget', {
+      serviceNamespace: appscaling.ServiceNamespace.ECS,
+      scalableDimension: 'ecs:service:DesiredCount',
+      minCapacity: props.minScalingCapacity || 0,
+      maxCapacity: props.maxScalingCapacity || 1,
+      resourceId: `service/${this.cluster.clusterName}/${this.service.serviceName}`
+    });
+
+    scalingTarget.scaleOnMetric('TotalMessagesScaling', {
+      metric: this.totalMessagesMetric,
+      scalingSteps: [
+        { lower: 0, upper: 0, change: -100 },
+        { lower: 1, change: 0 } // this is a bogus param - we require two for autoscaling
+      ],
+      evaluationPeriods: 1,
+      adjustmentType: appscaling.AdjustmentType.PERCENT_CHANGE_IN_CAPACITY
+    })
+
+    scalingTarget.scaleOnMetric('VisibleMessagesScaling', {
+      metric: this.sqsQueue.metricApproximateNumberOfMessagesVisible({
+        period: props.scaleUpAlarmInterval
+      }),
+      scalingSteps: [
+        { lower: 0, upper: 1, change: 0 },
+        { lower: 1, change: Math.round(Math.max(Math.min((props.maxScalingCapacity || 1) / 10, 100), 1)) }
+      ],
+      evaluationPeriods: 1
+    })
   }
 }